Move logic of openpmd-pipe to __main__.py
